♻️ Refactor run party yup schema
