+++ src/lib/types/api/validation/venogiaSubmission.ts
@@ -1,6 +1,7 @@
 import { type InferType, string, number, object, array, mixed, ObjectSchema } from 'yup';
 import { runSubmissionRequestSchema, type RunSubmissionRequest } from './runSubmission';
 import { yupRunTime } from './schemas/timeSchema';
+import { yupRunPartySchema } from './schemas/runRartySchema';
 
 const quest = ['venogia'];
 const categories = ['urgent_quest'];
@@ -12,6 +13,7 @@ export const venogiaRunSubmissionSchema: ObjectSchema<RunSubmissionRequest> =
 		quest: string().required().oneOf(quest),
 		rank: number().oneOf(ranks).required(),
 		category: string().required().oneOf(categories),
+		party: yupRunPartySchema(8),
 		time: yupRunTime(3600)
 	});
